+++ packages/gatsby/src/redux/index.js
@@ -66,6 +66,7 @@ const saveState = () => {
     `components`,
     `staticQueryComponents`,
     `webpackCompilationHash`,
+    `inferenceMetadata`,
   ])
 
   return writeToCache(pickedState)