+++ packages/gatsby/src/utils/page-data.js
@@ -1,7 +1,8 @@
 const fs = require(`fs-extra`)
 const path = require(`path`)
 const Promise = require(`bluebird`)
-const { chunk, isEqual } = require(`lodash`)
+const { isEqual } = require(`lodash`)
+const newPageKeys = []
 
 const getFilePath = ({ publicDir }, pagePath) => {
   const fixedPagePath = pagePath === `/` ? `index` : pagePath
@@ -13,36 +14,24 @@ const read = async ({ publicDir }, pagePath) => {
   return JSON.parse(rawPageData)
 }
 
-const write = async ({ publicDir }, page, result, webpackCompilationHash) => {
+const write = async ({ publicDir }, page, result) => {
   const filePath = getFilePath({ publicDir }, page.path)
   const body = {
     componentChunkName: page.componentChunkName,
     path: page.path,
     matchPath: page.matchPath,
-    webpackCompilationHash,
     result,
   }
   await fs.outputFile(filePath, JSON.stringify(body))
 }
 
-const updateCompilationHashes = (
-  { publicDir, workerPool },
-  pagePaths,
-  webpackCompilationHash
-) => {
-  const segments = chunk(pagePaths, 50)
-  return Promise.map(segments, segment =>
-    workerPool.updateCompilationHashes(
-      { publicDir },
-      segment,
-      webpackCompilationHash
-    )
-  )
-}
-
 const getNewPageKeys = (store, cacheData) =>
   new Promise(resolve => {
-    const newPageKeys = []
+    if (newPageKeys.length) {
+      resolve(newPageKeys)
+      return
+    }
+    console.log(cacheData.pages)
     if (cacheData.pages) {
       store.pages.forEach((value, key) => {
         if (!cacheData.pages.has(key)) {
@@ -80,7 +69,6 @@ const removePreviousPageData = (directory, store, cacheData) =>
 module.exports = {
   read,
   write,
-  updateCompilationHashes,
   getNewPageKeys,
   removePreviousPageData,
 }