+++ packages/gatsby-theme/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-workspace@1.0.3...gatsby-theme-workspace@1.0.4) (2019-11-10)
+
+**Note:** Version bump only for package gatsby-theme-workspace
+
+## [1.0.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-workspace@1.0.2...gatsby-theme-workspace@1.0.3) (2019-10-28)
+
+### Bug Fixes
+
+- update minor updates in packages ([#18875](https://github.com/gatsbyjs/gatsby/issues/18875)) ([b692879](https://github.com/gatsbyjs/gatsby/commit/b692879))
+
 ## [1.0.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-workspace@1.0.1...gatsby-theme-workspace@1.0.2) (2019-10-15)
 
 ### Bug Fixes